requirements.txt
newspaper3k
gTTS
pygame